fix(Switch): add defaultChecked
