+++ components/switch/index.js
@@ -1,71 +1,5 @@
-import React, { Component } from 'react'
-import PropTypes from 'prop-types'
-import classNames from 'classnames'
+import Switch from './Switch'
 import Provider from '../context'
 import './style/index'
 
-class Switch extends Component {
-  constructor (props) {
-    super(props)
-    this.state = {
-      checked: props.checked
-    }
-  }
-
-  componentWillReceiveProps (nextProps) {
-    if (nextProps.checked !== this.props.checked) {
-      this.setState({ checked: nextProps.checked }, () => {
-        this.props.onChange(nextProps.checked)
-      })
-    }
-  }
-
-  clickEvent () {
-    const { disabled, onChange, onClick } = this.props
-    if (!disabled) {
-      const res = onClick(this.state.checked)
-      if (typeof res === 'boolean' && !res) {
-        return
-      }
-      this.setState({ checked: !this.state.checked }, () => {
-        onChange(this.state.checked)
-      })
-    }
-  }
-
-  render () {
-    const { theme, content, disabled } = this.props
-    const { checked } = this.state
-    const sCls = classNames(
-      'theme__' + theme,
-      'hi-switch',
-      !checked && 'hi-switch--closed',
-      disabled && 'hi-switch--disabled'
-    )
-    return (
-      <span className={sCls} onClick={this.clickEvent.bind(this)}>
-        <span className='hi-switch__text'>
-          {content.length > 0 && (checked ? content[0] : content[1])}
-        </span>
-      </span>
-    )
-  }
-}
-
-Switch.defaultProps = {
-  content: [],
-  checked: false,
-  disabled: false,
-  onChange: () => {},
-  onClick: () => {
-    return true
-  }
-}
-Switch.propTypes = {
-  content: PropTypes.array,
-  checked: PropTypes.bool,
-  disabled: PropTypes.bool,
-  onChange: PropTypes.func,
-  onClick: PropTypes.func
-}
 export default Provider(Switch)
