+++ components/switch/style/index.scss
@@ -27,21 +27,28 @@ $basic-color: #4284f5 !default;
   cursor: pointer;
   position: relative;
   vertical-align: middle;
+  transition: 0.3s ease;
+  user-select: none;
 
   @include hi-switch-style();
 
-  &::after {
-    content: '';
-    width: 16px;
-    height: 16px;
+  &__dot {
+    $size: 16px;
+
+    width: $size;
+    height: $size;
     background: rgba(255, 255, 255, 1);
     position: absolute;
-    border-radius: 50%;
+    border-radius: $size / 2;
     top: 2px;
     left: 100%;
     transform: translateX(-100%);
     margin-left: -2px;
-    transition: all 200ms linear;
+    transition: 300ms ease;
+
+    .hi-switch:active & {
+      width: 22px;
+    }
   }
 
   &__text {
@@ -51,12 +58,12 @@ $basic-color: #4284f5 !default;
     top: 0;
     line-height: 20px;
     left: 6px;
-    transition: all 200ms linear;
+    transition: all 300ms ease;
     user-select: none;
   }
 
   &--closed {
-    &::after {
+    .hi-switch__dot {
       left: 2px;
       transform: translateX(2px);
     }
